style(SendProxy): match house structure for the managed side
Per maintainer's style request, align to sibling managers (TransmitManager/
ClientManager): move SendProxyValueKind to Sharp.Shared/Enums/, SendProxyValue +
SendProxyBatch to Sharp.Shared/Types/ (one interface per Managers file); nest
the callback as ISendProxyManager.DelegateSendProxyBatch (like
IClientManager.DelegateClientCommand); Native/Forward using-aliases; drop sealed;
EntityIndex unit type in the native bindings; multi-line doc comments. No
behavior change.
